urban-biker-lakehouse/requirements.txt
# Librerías necesarias (locales y para la App)
requests
pandas
pyarrow
fastparquet
snowflake-connector-python
python-dotenv
boto3
streamlit
pydeck
